MW-1536: Enhance TestUser class
+ Move creation of clients to the test user class, its knows best of the credentials
+ Made Config own class to user everywhere

+++ openexchange-test/src/com/openexchange/ajax/onboarding/tests/PlistSMSTest.java
@@ -126,8 +126,8 @@ public class PlistSMSTest extends AbstractPlistSMSTest {
     @TryAgain
     public void testDownload() throws Exception {
         PListDownloadTestHelper helper = new PListDownloadTestHelper(PlistSMSTest.class.getName());
-        int userId = testUser.getUserId().intValue();
-        int ctxId = testUser.getContextId().intValue();
+        int userId = testUser.getUserId();
+        int ctxId = testUser.getContextId();
         String hostname = AJAXConfig.getProperty(AJAXConfig.Property.HOSTNAME);
         String protocol = AJAXConfig.getProperty(AJAXConfig.Property.PROTOCOL);
         String url = getURL(userId, ctxId, "mailsync", "apple.iphone");
